refactor: extract disconnect logic
