+++ back/socket/game/index.js
@@ -1,12 +1,13 @@
 const { RoomManager } = require('../RoomManager');
-const sendGameImage = require('./gameImage');
+const { PRIVATE_ROOM_NAME } = require('../../config/roomConfig');
 const { enterRandom, enterPrivate } = require('./enterGame');
-const { sendMessage } = require('./message');
+const sendMessage = require('./message');
+const startPrivateGame = require('./startPrivateGame');
+const sendGameImage = require('./gameImage');
 const selectWord = require('./selectWord');
-const { PRIVATE_ROOM_NAME, escapeResultCode } = require('../../config/roomConfig');
-const { startPrivateGame } = require('./startPrivateGame');
 const exitRoom = require('./exitRoom');
 const changeRoomSetting = require('./changeRoomSetting');
+const disconnect = require('./disconnect');
 
 function setGameSocket(socket) {
   this.RoomManager = RoomManager;
@@ -36,38 +37,7 @@ function setGameSocket(socket) {
   socket.on('sendMessage', sendMessage.bind(this, gameSocket, roomInfo));
   socket.on('enterRandom', enterRandom.bind(this, gameSocket, roomInfo));
   socket.on('drawing', sendGameImage.bind(this, gameSocket, roomInfo));
-  socket.on('disconnect', () => {
-    const room = RoomManager.getRoomIfExist(roomInfo);
-    if (!room) return;
-
-    const userIdx = room.getUserIndexBySocketId(gameSocket);
-    if (userIdx < 0) return; // 방에서 유저 정보를 못 찾은경우 종료
-
-    // 방에 유저가 2명 이하면 방 정보 초기화
-    if (room.players.length <= 2) room.initRoomState();
-
-    // 방에 있는 플레이어 업데이트
-    room.removePlayer(userIdx);
-    gameSocket.leave();
-
-    const removeResult = room.getRoomStateAfterRemovePlayer(userIdx);
-    room.roomSettingAfterUserRemove(removeResult, gameSocket);
-    room.sendUserList(this.gameIo);
-
-    if (roomInfo.roomType === PRIVATE_ROOM_NAME) {
-      if (gameSocket.id === room.roomOwner) {
-        room.passRoomOwnerToNext();
-        room.sendUserList(this.gameIo);
-      }
-
-      if (room.players.length < 1) {
-        room.timer.stop();
-      }
-    } else if (room.players.length < 1) {
-      room.timer.stop();
-      RoomManager.deleteRoom(roomInfo.roomId);
-    }
-  });
+  socket.on('disconnect', () => disconnect(gameSocket, RoomManager, roomInfo));
 }
 
 module.exports = setGameSocket;
